+++ plugin-locationcomponent/src/test/java/com/mapbox/maps/plugin/locationcomponent/LocationProviderImplTest.kt
@@ -6,6 +6,7 @@ import com.mapbox.android.core.location.*
 import com.mapbox.android.core.permissions.PermissionsManager
 import com.mapbox.common.ShadowLogger
 import com.mapbox.geojson.Point
+import com.mapbox.maps.plugin.PuckBearingSource
 import io.mockk.*
 import org.junit.Assert.assertEquals
 import org.junit.Before
@@ -21,6 +22,7 @@ class LocationProviderImplTest {
   private val locationEngine = mockk<LocationEngine>(relaxed = true)
   private val locationConsumer1 = mockk<LocationConsumer>(relaxed = true)
   private val locationConsumer2 = mockk<LocationConsumer>(relaxed = true)
+  private val locationCompassEngine = mockk<LocationCompassEngine>(relaxed = true)
   private val locationEngineRequestSlot = CapturingSlot<LocationEngineRequest>()
   private val locationEngineCallbackSlot =
     CapturingSlot<LocationEngineCallback<LocationEngineResult>>()
@@ -31,7 +33,7 @@ class LocationProviderImplTest {
   fun setup() {
     mockkStatic(LocationEngineProvider::class)
     every { LocationEngineProvider.getBestLocationEngine(context) } returns locationEngine
-    locationProviderImpl = LocationProviderImpl(context)
+    locationProviderImpl = LocationProviderImpl(context, locationCompassEngine)
   }
 
   @Test
@@ -63,10 +65,14 @@ class LocationProviderImplTest {
     verify(exactly = 0) {
       locationEngine.getLastLocation(any())
     }
+    verify(exactly = 0) {
+      locationCompassEngine.addCompassListener(any())
+    }
   }
 
   @Test
   fun testAddLocationConsumerWithPermission() {
+    locationProviderImpl.updatePuckBearingSource(PuckBearingSource.HEADING)
     mockkStatic(PermissionsManager::class)
     every { PermissionsManager.areLocationPermissionsGranted(any()) } returns true
     locationProviderImpl.registerLocationConsumer(locationConsumer1)
@@ -92,10 +98,14 @@ class LocationProviderImplTest {
     verify(exactly = 1) {
       locationEngine.getLastLocation(any())
     }
+    verify(exactly = 1) {
+      locationCompassEngine.addCompassListener(any())
+    }
   }
 
   @Test
   fun testAddTwoLocationConsumer() {
+    locationProviderImpl.updatePuckBearingSource(PuckBearingSource.HEADING)
     locationProviderImpl.registerLocationConsumer(locationConsumer1)
     locationProviderImpl.registerLocationConsumer(locationConsumer2)
     verify(exactly = 1) {
@@ -108,6 +118,21 @@ class LocationProviderImplTest {
     verify(exactly = 2) {
       locationEngine.getLastLocation(any())
     }
+    verify(exactly = 1) {
+      locationCompassEngine.addCompassListener(any())
+    }
+  }
+
+  @Test
+  fun testUpdatePuckBearingSourceWithConsumer() {
+    locationProviderImpl.registerLocationConsumer(locationConsumer1)
+    verify(exactly = 0) {
+      locationCompassEngine.addCompassListener(any())
+    }
+    locationProviderImpl.updatePuckBearingSource(PuckBearingSource.HEADING)
+    verify(exactly = 1) {
+      locationCompassEngine.addCompassListener(any())
+    }
   }
 
   @Test
@@ -118,16 +143,23 @@ class LocationProviderImplTest {
     verify(exactly = 0) {
       locationEngine.removeLocationUpdates(any() as LocationEngineCallback<LocationEngineResult>)
     }
+    verify(exactly = 0) {
+      locationCompassEngine.addCompassListener(any())
+    }
   }
 
   @Test
   fun testRemoveAllLocationConsumer() {
+    locationProviderImpl.updatePuckBearingSource(PuckBearingSource.HEADING)
     locationProviderImpl.registerLocationConsumer(locationConsumer1)
     locationProviderImpl.registerLocationConsumer(locationConsumer2)
     locationProviderImpl.unRegisterLocationConsumer(locationConsumer2)
     locationProviderImpl.unRegisterLocationConsumer(locationConsumer1)
     verify(exactly = 1) {
       locationEngine.removeLocationUpdates(any() as LocationEngineCallback<LocationEngineResult>)
+    }
+    verify(exactly = 1) {
+      locationCompassEngine.addCompassListener(any())
     }
   }
 
@@ -153,6 +185,35 @@ class LocationProviderImplTest {
     verify { locationConsumer1.onLocationUpdated(Point.fromLngLat(12.0, 34.0)) }
     verify { locationConsumer1.onBearingUpdated(90.0) }
     verify { locationConsumer2.onLocationUpdated(Point.fromLngLat(12.0, 34.0)) }
+    verify { locationConsumer2.onBearingUpdated(90.0) }
+  }
+
+  @Test
+  fun testLocationUpdateWithCompass() {
+    locationProviderImpl.updatePuckBearingSource(PuckBearingSource.HEADING)
+    val locationEngineResult = mockk<LocationEngineResult>(relaxed = true)
+    val location = mockk<Location>(relaxed = true)
+    every { locationEngineResult.lastLocation } returns location
+    every { location.longitude } returns 12.0
+    every { location.latitude } returns 34.0
+    every { location.bearing } returns 90.0f
+
+    locationProviderImpl.registerLocationConsumer(locationConsumer1)
+    locationProviderImpl.registerLocationConsumer(locationConsumer2)
+    verify(exactly = 1) {
+      locationEngine.requestLocationUpdates(
+        any(),
+        capture(locationEngineCallbackSlot),
+        any()
+      )
+    }
+    locationEngineCallbackSlot.captured.onSuccess(locationEngineResult)
+    verify { locationConsumer1.onLocationUpdated(Point.fromLngLat(12.0, 34.0)) }
+    verify(exactly = 0) { locationConsumer1.onBearingUpdated(90.0) }
+    verify { locationConsumer2.onLocationUpdated(Point.fromLngLat(12.0, 34.0)) }
+    verify(exactly = 0) { locationConsumer2.onBearingUpdated(90.0) }
+    locationProviderImpl.onCompassChanged(90.0f)
+    verify { locationConsumer1.onBearingUpdated(90.0) }
     verify { locationConsumer2.onBearingUpdated(90.0) }
   }
 }